+++ test/wdio/stencil.config.ts
@@ -9,6 +9,10 @@ export const config: Config = {
     {
       type: 'dist',
     },
+    {
+      type: 'dist-custom-elements',
+      dir: 'test-components',
+    },
   ],
   plugins: [sass()],
   buildDist: true,